requirements.txt
google-auth==2.21.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.39.0
numpy==2.1.2
pandas==1.5.3
